Installation does not correctly publish BizFx files
During installation, a task copies the out of the box BizFx files into
the publish directory. The task was not copying the subfolder structure
and was leaving out the important "assets" folder.

Modified the script to -Recurse but to also delete the destination
folder prior to publishing (if it already existed)